wip: load board with automatic size and position for non base layouts
